% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption_pop.R
\name{caption_pop}
\alias{caption_pop}
\title{Add Custom Captions with Icons to a geom_pop object}
\usage{
caption_pop(caption_size = 1, icon_size = 1, hjust = 0.5, text = NULL)
}
\arguments{
\item{caption_size}{Numeric. The font size of the caption text. Default is \code{3}.}

\item{icon_size}{Numeric. The width of the images (icons) in pixels within the caption. Default is \code{20}.}

\item{hjust}{Numeric. The horizontal justification of the caption. Values range from \code{0} (left) to \code{1} (right).
Default is \code{0.5}.}

\item{text}{Named list or named vector. Custom text descriptions for each icon. The names should correspond
to the \code{image} identifiers in the plot data. If \code{NULL}, defaults to \code{"persons"}.}
}
\description{
The \code{caption_pop} function allows you to add custom captions to a \code{geom_pop} object.
The function generates a caption based on the images used in the plot.
The caption includes the number of individuals represented by each image.
The function also allows you to specify the size of the caption text and
the icons used in the caption.
}
\examples{
# Example usage:
ggplot() +
 geom_pop(data = df_iris_prop, aes(icon = icon, group=type, color=type),
        size = 1.3, arrange = T) +
        theme_void() +
        caption_pop(caption_size = 1, icon_size = 1, hjust=.8)
        
}
