% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key.R
\name{draw_key_pop_image}
\alias{draw_key_pop_image}
\title{Key drawing function for population-based image keys}
\usage{
draw_key_pop_image(data, params, size)
}
\arguments{
\item{data}{A data frame containing the scaled aesthetics for the key.
It must include a \code{colour} column for color, an \code{alpha} column for transparency,
and an \code{icon} column with the names of the icon files (without extension) to be used.}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{The width and height of the key in mm. This value is not used directly in this function.}
}
\value{
A grid grob containing the image icons with the specified colors and transparency.
}
\description{
This function creates a custom key for displaying population-based image icons
in the legend of a ggplot2 plot. Each group can be assigned a different icon
based on the information in the \code{data$icon} column, and the icons can be colorized
according to the specified \code{colour} and \code{alpha} aesthetics.
}
\details{
This function relies on \code{ggimage:::color_image} and \code{ggplot2:::ggname}, which are internal functions.
Their use is necessary for correct functionality, and no exported alternatives exist.
We acknowledge the potential risks associated with \code{:::} usage, but at present, these functions
provide essential behavior for rendering images within ggplot2.
}
