% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_df_coordinates.R
\name{fetch_df_coordinates}
\alias{fetch_df_coordinates}
\title{Fetches the \code{df_coordinates_final} Dataset}
\usage{
fetch_df_coordinates()
}
\value{
A data frame containing the \code{df_coordinates_final} dataset.
}
\description{
Downloads and caches the \code{df_coordinates_final} dataset if it is not already cached locally.
This function ensures that the dataset is downloaded only once and loaded into memory
without cluttering the global environment. The dataset is stored in a package-specific
cache directory and retrieved efficiently for subsequent uses.
}
\details{
The dataset is downloaded from GitHub
The file is cached in a directory specific to the package, which is determined
using \code{\link[tools]{R_user_dir}}. If the dataset is already cached, it will
be loaded directly from the cache instead of downloading again.
}
\examples{
\donttest{
df <- fetch_df_coordinates()
head(df)
}

}
