% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pop.R
\name{geom_pop}
\alias{geom_pop}
\title{Create a circular representative population chart
#'
Draws a circular representative population chart based on the proportion of the groups,
where each point (person) represents a determined number of individuals.
Every person is represented by an image with a given icon.}
\usage{
geom_pop(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  icon = "ggmale",
  group_var = NULL,
  sample_size = NULL,
  arrange = FALSE,
  sum_var = NULL,
  facet = NULL,
  size = 3,
  dpi = 50,
  legend_icons = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this layer.
When using a \verb{geom_*()} function to construct a layer, the \code{stat}
argument can be used to override the default coupling between geoms and
stats. The \code{stat} argument accepts the following:
\itemize{
\item A \code{Stat} ggproto subclass, for example \code{StatCount}.
\item A string naming the stat. To give the stat as a string, strip the
function name of the \code{stat_} prefix. For example, to use \code{stat_count()},
give the stat as \code{"count"}.
\item For more information and other ways to specify the stat, see the
\link[ggplot2:layer_stats]{layer stat} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{na.rm}{logical, whether remove NA values}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}

\item{icon}{The icon to be used in the chart.}

\item{group_var}{The variable used to group individuals.}

\item{sample_size}{The total number of individuals (points) to be drawn.}

\item{arrange}{Logical; if TRUE, the output data is arranged by group.}

\item{sum_var}{Optional variable to sum over instead of counting.}

\item{facet}{Optional facetting variable.}

\item{size}{The size of the points.}

\item{dpi}{Height (in \strong{pixels}) of the PNG icon when rendered with \code{fontawesome::fa_png()}.
Higher values produce sharper icons. Defaults to 50. This affects \strong{image dpi}, not icon size in the plot.}

\item{legend_icons}{Logical; if TRUE, the legend will display the selected icons by the user.}

\item{...}{additional parameters}
}
\value{
A ggplot object with a circular representative population chart.
}
\description{
Create a circular representative population chart
#'
Draws a circular representative population chart based on the proportion of the groups,
where each point (person) represents a determined number of individuals.
Every person is represented by an image with a given icon.
}
\section{Aesthetics}{

geom_pop employs the following aesthetics:
\itemize{
\item \strong{sample_size} - The number of individuals to be represented in the chart.
\item \strong{alpha} - The transparency of the points.
\item \strong{color} - The color of the points.
\item \strong{size} - The size of the points.
}
}

