% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_legend_icon.R
\name{ggplot_add.ggpop_legend_icon}
\alias{ggplot_add.ggpop_legend_icon}
\title{S3 method for adding ggpop_legend_icon objects to ggplot}
\usage{
\method{ggplot_add}{ggpop_legend_icon}(object, plot, object_name)
}
\description{
S3 method for adding ggpop_legend_icon objects to ggplot
}
