% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpop-package.R
\docType{package}
\name{ggpop-package}
\alias{ggpop}
\alias{ggpop-package}
\title{ggpop: Visualizing Population Data (ggpop)}
\description{
ggpop is an R package that creates visually engaging and informative population charts. ggpop allows users to represent population data proportionally using customizable icons, facilitating the creation of circular representative population charts.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jurjoroa/ggpop/issues}
}

}
\author{
\strong{Maintainer}: Jorge Roa \email{jorgeroa@stanford.edu} (\href{https://orcid.org/0000-0002-3972-9793}{ORCID})

Authors:
\itemize{
  \item Carlos Pineda-Antunez \email{cpinedaa@uw.edu} (\href{https://orcid.org/0000-0002-8352-7080}{ORCID})
  \item Fernando Alarid-Escudero \email{falarid@stanford.edu} (\href{https://orcid.org/0000-0001-5076-1172}{ORCID})
}

}
\keyword{internal}
