% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icon-utils.R
\name{ggpop_markers}
\alias{ggpop_markers}
\title{List the icon markers ggpop can render by name}
\usage{
ggpop_markers(icon_path = getOption("ggpop.icon_path"))
}
\arguments{
\item{icon_path}{Optional path to a folder of user SVG icons. Defaults to
\code{getOption("ggpop.icon_path")}.}
}
\value{
A list with element \code{bundled} (character vector of marker names)
and, when \code{icon_path} resolves to a directory, \code{user}.
}
\description{
Returns the bundled ggpop marker names and, if an icon directory is given,
the names of the user SVGs found there. These names (plus any Font Awesome
name) are valid values for the \code{icon} aesthetic of \code{geom_pop()} and
\code{geom_icon_point()}.
}
\examples{
ggpop_markers()

}
