% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key.R
\name{key_glyph_icon_point}
\alias{key_glyph_icon_point}
\title{Key glyph for icon point legend}
\usage{
key_glyph_icon_point(key_data, params, size)
}
\description{
Key glyph for icon point legend
}
\keyword{internal}
