% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process Population Data for Visualization}
\usage{
process_data(
  data,
  high_group_var = NULL,
  group_var = NULL,
  sum_var = NULL,
  sample_size = 1000,
  replace = TRUE
)
}
\arguments{
\item{data}{A data frame containing the population data to be processed.}

\item{high_group_var}{Character vector, optional. The variables used to group individuals hierarchically. This should be a categorical variable. If provided, the function samples individuals within each group defined by these variables.}

\item{group_var}{Quosure. The variable used to group individuals in the dataset. This should be a categorical variable.}

\item{sum_var}{Quosure, optional. The variable to sum over within each group. If \code{NULL}, the function counts the number of individuals per group.}

\item{sample_size}{Integer. The total number of individuals to sample based on group proportions. Must be a positive integer.}
}
\value{
A tibble (data frame) with the following columns:
\describe{
\item{type}{The sampled group type.}
\item{pos}{The position index of the sampled individual.}
\item{group}{The group identifier.}
\item{n}{The count of individuals in the group.}
\item{prop}{The proportion of the group relative to the total population.}
}
}
\description{
The \code{process_data} function processes a dataset to calculate group proportions and generates a sampled dataset based on specified parameters. This processed data is suitable for creating visual representations, such as population charts, where each sample represents a group with associated counts and proportions.
}
