% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_legend_icon.R
\name{scale_legend_icon}
\alias{scale_legend_icon}
\title{Legend helper for geom_pop icon legends}
\usage{
scale_legend_icon(
  size = 10,
  margin = NULL,
  size_default = 10,
  margin_default = ggplot2::margin(0, 0, 0, 0),
  unit = "mm",
  ...
)
}
\arguments{
\item{size}{Numeric. Legend key size (default 10).}

\item{margin}{ggplot2::margin() for plot margin (default margin_default).}

\item{size_default}{Numeric. Fallback size used when size is invalid (default 10).}

\item{margin_default}{ggplot2::margin() used when margin is NULL.}

\item{unit}{Character. Unit for legend key sizing (default "mm").}

\item{...}{Additional arguments forwarded to ggplot2::guide_legend()}
}
\description{
Legend helper for geom_pop icon legends
}
