% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpop-package.R
\docType{package}
\name{ggpop-package}
\alias{ggpop}
\alias{ggpop-package}
\title{ggpop: ggpop - R package to visualize population data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create in R visually engaging and informative population charts. ggpop allows users to represent population data proportionally using customizable icons, facilitating the creation of circular representative population charts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jurjoroa.github.io/ggpop/}
  \item Report bugs at \url{https://github.com/jurjoroa/ggpop/issues}
}

}
\author{
\strong{Maintainer}: Jorge A. Roa-Contreras \email{jorgeroa@stanford.edu} (\href{https://orcid.org/0000-0002-3972-9793}{ORCID})

Authors:
\itemize{
  \item Ralitza Soultanova \email{Ralitza.soultanova@gmail.com} (\href{https://orcid.org/0009-0000-9324-5653}{ORCID})
  \item Fernando Alarid-Escudero \email{falarid@stanford.edu} (\href{https://orcid.org/0000-0001-5076-1172}{ORCID})
  \item Carlos Pineda-Antunez \email{cpinedaa@uw.edu} (\href{https://orcid.org/0000-0002-8352-7080}{ORCID})
}

}
\keyword{internal}
