% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpop-package.R
\docType{package}
\name{ggpop-package}
\alias{ggpop}
\alias{ggpop-package}
\title{ggpop: Icon-Based Population Charts for R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right; width: 120px;'}}

\code{ggpop} is a \code{ggplot2} extension for creating icon-based population charts
and pictogram plots. Use \code{\link[=geom_pop]{geom_pop()}} and \code{\link[=geom_icon_point]{geom_icon_point()}} to visualize
proportion and population data with 2,000+ Font Awesome icons.
}
\section{Main functions}{

\itemize{
\item \code{\link[=geom_pop]{geom_pop()}} -- proportional icon grids
\item \code{\link[=geom_icon_point]{geom_icon_point()}} -- icon scatter plots
\item \code{\link[=process_data]{process_data()}} -- prepare count data for plotting
\item \code{\link[=fa_icons]{fa_icons()}} -- search Font Awesome icon names
\item \code{\link[=theme_pop]{theme_pop()}}, \code{\link[=theme_pop_dark]{theme_pop_dark()}}, \code{\link[=theme_pop_minimal]{theme_pop_minimal()}} -- built-in themes
}
}

\section{process_data()}{

Converts count data to one row per icon. \code{group_var} and \code{sum_var} are
unquoted; \code{high_group_var} takes a character string for faceted charts.
\preformatted{
df_plot <- process_data(
  data        = data.frame(sex = c("Female", "Male"), n = c(55, 45)),
  group_var   = sex,
  sum_var     = n,
  sample_size = 20
)
}
}

\section{geom_pop()}{

Draws icon grids. Add an \code{icon} column, map \code{icon} and \code{color} in \code{aes()}.
Do not map \code{x} or \code{y}.
\preformatted{
ggplot() +
  geom_pop(data = df_plot, aes(icon = icon, color = type), size = 2) +
  scale_color_manual(values = c(Female = "#C0392B", Male = "#2980B9")) +
  theme_pop()
}
}

\section{geom_icon_point()}{

Drop-in replacement for \code{geom_point()} using Font Awesome icons.
\preformatted{
ggplot(iris, aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
  geom_icon_point(icon = "seedling", size = 1)
}
}

\section{fa_icons()}{

Search the bundled Font Awesome icon list by keyword.
\preformatted{
fa_icons(query = "person")
}
}

\section{Themes}{

Three built-in themes optimized for icon charts:
\code{theme_pop()}, \code{theme_pop_dark()}, \code{theme_pop_minimal()}.
}

\examples{
library(ggplot2)
library(dplyr)

## -------------------------------------------------------
## geom_pop(): population icon grid
## -------------------------------------------------------
df_plot <- process_data(
  data        = data.frame(sex = c("Female", "Male"), n = c(55, 45)),
  group_var   = sex,
  sum_var     = n,
  sample_size = 20
) \%>\%
  mutate(icon = ifelse(type == "Female", "person-dress", "person"))

ggplot() +
  geom_pop(data = df_plot, aes(icon = icon, color = type), size = 2) +
  scale_color_manual(values = c(Female = "#C0392B", Male = "#2980B9")) +
  theme_pop() +
  labs(title = "Population by sex", color = NULL)

## -------------------------------------------------------
## geom_icon_point(): icon scatter plot
## -------------------------------------------------------
ggplot(iris, aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
  geom_icon_point(icon = "seedling", size = 1) +
  scale_color_manual(values = c(
    setosa     = "#43A047",
    versicolor = "#1E88E5",
    virginica  = "#E53935"
  )) +
  labs(title = "Iris dataset", x = "Sepal Length", y = "Petal Length")

}
\seealso{
Useful links:
\itemize{
  \item \url{https://jurjoroa.github.io/ggpop/}
  \item Report bugs at \url{https://github.com/jurjoroa/ggpop/issues}
}

}
\author{
\strong{Maintainer}: Jorge A. Roa-Contreras \email{jorgeroa@stanford.edu} (\href{https://orcid.org/0000-0002-3972-9793}{ORCID})

Authors:
\itemize{
  \item Ralitza Soultanova \email{Ralitza.soultanova@gmail.com} (\href{https://orcid.org/0009-0000-9324-5653}{ORCID})
  \item Fernando Alarid-Escudero \email{falarid@stanford.edu} (\href{https://orcid.org/0000-0001-5076-1172}{ORCID})
  \item Carlos Pineda-Antunez \email{cpinedaa@uw.edu} (\href{https://orcid.org/0000-0002-8352-7080}{ORCID})
}

}
