% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_legend_icon.R
\name{scale_legend_icon}
\alias{scale_legend_icon}
\title{Legend helper for geom_pop icon legends}
\usage{
scale_legend_icon(size = 10, margin = NULL, ...)
}
\arguments{
\item{size}{Numeric. Legend key size in mm (this controls icon size).}

\item{margin}{ggplot2::margin() for plot margin (defaults to bottom padding).}

\item{...}{Additional arguments forwarded to ggplot2::guide_legend()}
}
\description{
Legend helper for geom_pop icon legends
}
